Extract apply_control_update into agent/document_view/apply.rs
